upgrade socket module
