+++ Blik_2024_room.js
@@ -1,0 +1,112 @@
+ import {document,observe as activate} from "./Blik_2023_fragment.js";
+ import {note,infer,compose,observe,buffer,wait,either,drop,tether} from "./Blik_2023_inference.js";
+ import {fetch,resolve} from "./Blik_2023_interface.js";
+ import {merge} from "./Blik_2023_search.js";
+
+ export async function open(server,room="./Blik_2020_room.js")
+{let {default:{WebSocketServer}}=await import("./einaros_2011_ws.js");
+ let connection=buffer(async function connection(host,peer,request)
+{let {referer,cookie}=request.headers;
+ let author=cookie&&cookie.match(/signature=[^;]+/);
+ let url=author&&author[0].replace("=","/");
+ ({author}=url?await fetch(url):{author:"anonymous"});
+ activate.call(Object.assign(peer,{author}),{message:infer(transmit,host)});
+},note.bind(1));
+ return compose([{server}],Reflect.construct,{rooms:{}},Object.assign,{connection},tether(observe))(WebSocketServer);
+};
+
+ export var transmit=either
+(function(peer,host,event)
+{let message=JSON.parse(event.data);
+ respond[message.type||"message"]?.call(host,message,peer);
+ return true;
+},compose(drop(-1,1),(peer,error)=>
+ peer.send(JSON.stringify({error:note.call(1,error).message})))
+);
+
+ var respond=
+ {join({room},peer)
+{this.rooms[room]=this.rooms[room]||{messages:[]};
+ let {author}=merge(peer,{rooms:new Set([room])},0);
+ let message=JSON.stringify({author:system,message:author+" joined "+room,room});
+ this.clients.forEach(client=>client!==peer&&client.rooms.has(room)&&client.send(message));
+ let {length}=this.rooms[room].messages;
+ if(length)
+ peer.send(JSON.stringify({type:"history",history:this.rooms[room].messages}));
+},sign({author},peer){peer.author=author;}
+ ,signal({room},peer)
+{let response=JSON.stringify({type:"signal",author:peer.author});
+ this.clients.forEach(client=>client.readyState===1&&
+ client.send(response));
+},message({message},{author})
+{let response=JSON.stringify({author,message});
+ this.clients.forEach(client=>client.readyState===1&&
+ client.send(response));
+},save:async function({room,updates,version})
+{this.room=this.adapter.rooms[room]||this.join(room);
+ let {EditorState,collab,receiveUpdates,getSyncedVersion,ChangeSet}=
+ await import("./haverbeke_2020_codemirror.js");
+ if(!this.room.track)this.room.track=await fetch(room).then(doc=>
+ EditorState.create({doc:doc.toString(),extensions:[collab()]}));
+ //if(note([version,getSyncedVersion(this.room.content)]).reduce((next,past)=>next>past))
+ await update(this.room.track,updates);
+ let body=this.room.track.doc.toString();
+ let headers={"Content-Type":"text/plain"};
+ let message=await fetch(room+"?force=overwrite",{method:"put",body,headers});
+ if(message instanceof Error)
+ return this.emit("message",{author:system,message});
+ else message=this.author+" edited "+room;
+[["save",{author:this.author,room,updates}]
+,["message",{author:system,message}]
+].map(event=>
+ this.server.sockets.in(room).emit(...event));
+},disconnect(text)
+{Object.keys(this.adapter.rooms).forEach(room=>
+ this.leave(room)&&
+ this._events.message.bind(this)({author:system,message:this.author+" left "+room,room}));
+},put:async function({room,body})
+{this.room=this.adapter.rooms[room]||this.join(room);
+ if(!this.room.content)
+ await fetch(room).then(content=>this.room.content=content);
+ this.room.content=JSON.stringify(body);
+ let message=await fetch(room+"?force=overwrite"
+,{method:"put",body:this.room.content
+ ,headers:{"Content-Type":"text/plain"}
+ });
+ if(message instanceof Error)
+ return this.emit("message",{author:system,message});
+[["put",{body}]
+,["message",{author:system,message:this.author+" updated "+room}]
+].map(event=>this.server.sockets.in(room).emit(...event))
+}};
+
+ export var receipt=
+ {message({type,message,author:{name,image}})
+{message=document({li:{img:{src:image||"svg/deer/document",height:"12px"},div:{"#text":name||"anonymous"},span:{"#text":message}}});
+ note(arguments[0]);
+ if(name==="system")
+ compose(wait(10000),{style:"opacity:0;transition:all 1s;"},Object.assign,wait(1000),"remove")(message);
+ this.message.parentNode.querySelector("ul").append(message);
+},signal({author})
+{let label=this.message.parentNode;
+ let list=label.querySelector("ul")||label.appendChild(document({ul:{}}));
+ let node=list.querySelector("span#signal");
+ let entry=document({span:{id:"signal",signal:window.document.createRange().createContextualFragment(author+" is typing...")}});
+ list[node?"replaceChild":"appendChild"](entry,node);
+ compose(wait(3000),node=>node.parentNode&&node.remove())(entry);
+},save({author,room,updates})
+{if(author==window.subject.labels.message)
+ return window.Tone.Transport.start();
+ update(this.room.content.viewState.state,updates).then(updates=>
+ this.room.content.update([updates]));
+}};
+
+ var update=(state,updates)=>
+ import("./haverbeke_2020_codemirror.js").then(({ChangeSet,receiveUpdates})=>
+ receiveUpdates(state,updates.map(({changes})=>
+ Array.isArray(changes)
+?ChangeSet.fromJSON(changes)
+:changes.toJSON()).map((changes,index)=>
+ Object.assign(updates[index],{changes}))));
+
+ var system={name:"system"};
